+++ packages/jest-haste-map/src/worker.js
@@ -77,7 +77,13 @@ export async function worker(data: WorkerMessage): Promise<WorkerMetadata> {
     }
 
     if (computeDependencies) {
-      dependencies = extractRequires(getContent());
+      const content = getContent();
+      dependencies = Array.from(
+        data.dependencyExtractor
+          ? // $FlowFixMe
+            require(data.dependencyExtractor)(content, extractRequires)
+          : extractRequires(content),
+      );
     }
 
     if (id) {